last update before sleep
